// src/app/models/empresa.model.ts
export interface Empresa {
    id_empresa: string; // ID de la empresa
    nombre: string; // Nombre de la empresa
    descripcion: string; // Descripción de la empresa
    direccion: string; // Dirección de la empresa
    telefono: string; // Teléfono de la empresa
    correoElectronico: string; // Correo electrónico de la empresa
    sitioWeb: string; // Sitio web de la empresa
    sector: string; // Sector o industria de la empresa
    fechaFundacion: Date; // Fecha de fundación de la empresa
    logoUrl: string; // URL de la imagen del logo de la empresa
}